코드 구조 정리를 위한 함수 위치 조정
비밀번호 관련 유효성 검사 함수를 validatePassword.js로 이동
이메일 관련 유효성 검사 함수를 validateEmail.js로 통합
