+++ javascript/functions/vaildatePassword.js
@@ -1,6 +1,6 @@
 import { makeWarningSpanTag, changeRedBorder } from "./handleSpanTag.js";
-import { $passwordInput } from "../tags.js";
-export { addWarningPasswordMsg };
+import { $passwordInput, $repasswordInput } from "../tags.js";
+export { addWarningPasswordMsg, vaildateRepassword, vaildatePassword };
 
 function addWarningPasswordMsg() {
   if (!$passwordInput.value) {
@@ -10,5 +10,28 @@ function addWarningPasswordMsg() {
       $passwordInput,
       "passwordSpan"
     );
+  }
+}
+
+function vaildateRepassword() {
+  const firstPassword = $passwordInput.value;
+  const secondPassword = $repasswordInput.value;
+  if (firstPassword === secondPassword) {
+    return true;
+  } else {
+    return false;
+  }
+}
+
+function vaildatePassword() {
+  const passwordValue = $passwordInput.value;
+  if (
+    passwordValue.length < 8 ||
+    /^\d+$/.test(passwordValue) ||
+    /^[a-zA-Z]+$/.test(passwordValue)
+  ) {
+    return false;
+  } else {
+    return true;
   }
 }
